Horrible hack to allow symfony/http-foundation >= 6.0 (definitely need a better solution)
