+++ src/Ratchet/Session/Storage/Proxy/VirtualProxy.php
@@ -1,54 +1,7 @@
 <?php
-namespace Ratchet\Session\Storage\Proxy;
-use Symfony\Component\HttpFoundation\Session\Storage\Proxy\SessionHandlerProxy;
 
-class VirtualProxy extends SessionHandlerProxy {
-    /**
-     * @var string
-     */
-    protected $_sessionId;
-
-    /**
-     * @var string
-     */
-    protected $_sessionName;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function __construct(\SessionHandlerInterface $handler) {
-        parent::__construct($handler);
-
-        $this->saveHandlerName = 'user';
-        $this->_sessionName    = ini_get('session.name');
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getId() {
-        return $this->_sessionId;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setId($id) {
-        $this->_sessionId = $id;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName() {
-        return $this->_sessionName;
-    }
-
-    /**
-     * DO NOT CALL THIS METHOD
-     * @internal
-     */
-    public function setName($name) {
-        throw new \RuntimeException("Can not change session name in VirtualProxy");
-    }
+if (version_compare(\Composer\InstalledVersions::getVersion('symfony/http-foundation'), '6.0.0') === -1) {
+    include 'VirtualProxy<8.php';
+} else {
+    include 'VirtualProxy8+.php';
 }
